Rename boardState
