+++ src/minesweeper/board/boardStore.svelte.ts
